requirements.txt
Flask
numpy
opencv-python
matplotlib
ultralytics
scikit-learn
pyyaml
pandas
gdown